+++ src/constants.ts
@@ -1,9 +1,26 @@
 import { createSelector } from '@volkovlabs/jest-selectors';
 
-/**
- * All Constants here
- */
+import { ColumnAlignment } from '@/types';
 
+/**
+ * Default Column Appearance
+ */
+export const DEFAULT_COLUMN_APPEARANCE = {
+  wrap: true,
+  alignment: ColumnAlignment.START,
+  width: {
+    auto: true,
+    min: 20,
+    value: 100,
+  },
+  background: {
+    applyToRow: false,
+  },
+};
+
+/**
+ * Test Identifiers
+ */
 export const TEST_IDS = {
   panel: {
     root: createSelector('data-testid panel'),
@@ -40,6 +57,15 @@ export const TEST_IDS = {
     fieldSortEnabled: createSelector('data-testid column-editor field-sort-enabled'),
     fieldAppearanceBackgroundApplyToRow: createSelector(
       'data-testid column-editor field-appearance-background-apply-to-row'
+    ),
+    fieldAppearanceWidthAuto: createSelector('data-testid column-editor field-appearance-width-auto'),
+    fieldAppearanceWidthMin: createSelector('data-testid column-editor field-appearance-width-min'),
+    fieldAppearanceWidthMax: createSelector('data-testid column-editor field-appearance-width-max'),
+    fieldAppearanceWidthValue: createSelector('data-testid column-editor field-appearance-width-value'),
+    fieldAppearanceWrap: createSelector('data-testid column-editor field-appearance-wrap'),
+    fieldAppearanceAlignment: createSelector('data-testid column-editor field-appearance-alignment'),
+    fieldAppearanceAlignmentOption: createSelector(
+      (name: unknown) => `wcolumn-editor field-appearance-alignment-option-${name}`
     ),
   },
   defaultCellRenderer: {